add mine alert interface attributes
